# web/public/docs/api/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "DIY Accounting Submit API",
    "description" : "DIY Accounting Submit API documentation",
    "version" : "1.0.0",
    "termsOfService" : "https://submit.diyaccounting.co.uk/terms.html",
    "contact" : {
      "email" : "admin@diyaccounting.co.uk"
    },
    "license" : {
      "name" : "PolyForm Internal Use License 1.0.0",
      "url" : "https://github.com/diy-accounting-uk/submit.diyaccounting.co.uk/blob/main/LICENSE"
    },
    "x-authentication-guide" : {
      "description" : "This API is protected by Amazon Cognito. To use the API:\n\n1. **Via Website**: If you are logged into the DIY Accounting Submit website, you can access the Swagger UI directly and API calls will use your existing session cookies.\n\n2. **External Access**: \n   - First complete the Cognito hosted UI OAuth flow to get an authorization code\n   - Exchange that code for an access token by calling `/api/v1/cognito/token`\n   - Include the token in the `Authorization: Bearer <token>` header\n\n3. **For HMRC Integration**: Some endpoints require HMRC authorization tokens obtained via `/api/v1/hmrc/token`\n\n**Note**: All endpoints require proper CORS headers and are accessible from the same domain as the main application."
    }
  },
  "servers" : [ {
    "url" : "${env.DIY_SUBMIT_APEX_URL}api/v1/",
    "description" : "DIY Accounting Submit API documentation"
  } ],
  "paths" : {
...
    "/companies-house/company/{companyNumber}" : {
      "get" : {
        "summary" : "Get a Companies House company profile",
        "description" : "Retrieves the Companies House profile for a company number",
        "operationId" : "getCompanyProfile",
        "parameters" : [ {
          "name" : "companyNumber",
          "in" : "path",
          "required" : true,
          "description" : "The 8-character company number",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
    "/companies-house/search" : {
      "get" : {
        "summary" : "Search Companies House",
        "description" : "Searches Companies House for a company by name or number",
        "operationId" : "searchCompanies",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : true,
          "description" : "Company name or number search term",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "itemsPerPage",
          "in" : "query",
          "required" : false,
          "description" : "Results per page (1-50, default 20)",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "startIndex",
          "in" : "query",
          "required" : false,
          "description" : "Zero-based offset into the result set",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
    "/companies-house/company/{companyNumber}/registered-office-address" : {
      "get" : {
        "summary" : "Get a company's registered office address",
        "description" : "Reads the current registered office address from the public register, with its etag",
        "operationId" : "getCompaniesHouseRegisteredOfficeAddress",
        "parameters" : [ {
          "name" : "companyNumber",
          "in" : "path",
          "required" : true,
          "description" : "The 8-character company number",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "CognitoAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Cognito JWT token"
      },
      "HmrcAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "description" : "HMRC OAuth2 token"
      }
    }
  }
}